Split up the descriptor::indexes::species module
