+++ rascaline/src/descriptor/indexes/species/mod.rs
@@ -1,0 +1,8 @@
+mod structure;
+pub use self::structure::StructureSpeciesSamples;
+
+mod atom;
+pub use self::atom::AtomSpeciesSamples;
+
+mod three_bodies;
+pub use self::three_bodies::ThreeBodiesSpeciesSamples;
